[REL] Migration product_unique to 11.0 * [FIX] Update Readme * [REM] barcode contrains becase odoo incorporates that * [FIX] change UserError to ValidationError * [MOV] change product.py file name and put it into models folder
