+++ product_unique/README.rst
@@ -1,6 +1,14 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-  :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-  :alt: License: AGPL-3
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
 
 ==============
 Product Unique
@@ -15,15 +23,12 @@ To install this module, you need to:
 
 #. Just install this module.
 
-
 Configuration
 =============
 
 To configure this module, you need to:
 
 #. No configuration nedeed.
-
-
 
 Usage
 =====
@@ -32,14 +37,9 @@ To use this module, you need to:
 
 #. Set an default code and barcode to the product.
 
-
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-  :alt: Try me on Runbot
-  :target: https://runbot.adhoc.com.ar/
-
-.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
-.. branch is "9.0" for example
-
+   :alt: Try me on Runbot
+   :target: http://runbot.adhoc.com.ar/
 
 Bug Tracker
 ===========
@@ -55,19 +55,16 @@ Credits
 Images
 ------
 
-* ADHOC SA: `Icon <http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png>`_.
+* |company| |icon|
 
 Contributors
 ------------
 
-
 Maintainer
 ----------
 
-.. image:: http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png
-  :alt: Odoo Community Association
-  :target: https://www.adhoc.com.ar
+|company_logo|
 
-This module is maintained by the ADHOC SA.
+This module is maintained by the |company|.
 
 To contribute to this module, please visit https://www.adhoc.com.ar.
