+++ product_unique/models/__init__.py
@@ -3,4 +3,4 @@
 # directory
 ##############################################################################
 
-from . import models
+from . import product_product
